Gemini_on_Googlechat/requirements.txt
functions-framework==3.*
Flask==2.2.5
google-cloud-discoveryengine==0.11.4
google-api-python-client==2.86.0
google-auth>=2.17.3
google-auth-oauthlib==1.0.0
protobuf>=4.22.3,<5.0.0dev
vertexai>=1.43.0